Move traits.hpp to top level (i.e. public) directory.
